Better high concurrency handling via "add".
The previous method of testing if a lock existed, and if not existing
add the lock reduced repeated creations but did not eliminate them.

The best way to create a custom lock key in memcache is to use the "add"
method instead of the "set" method. The add method will return false if
the key already exists. Thus, if only add is used, we should be able to
handle increased concurrency much better.

memcache_store and dalli_store both accept the :unless_exist option
which uses the "add" method if set to true. This commit uses this
option, which eliminates repeated test creation in my tests.