Fix conversion scoring when test has spaces
Participating tests was stored in the cache with an array of test names,
and the test names were unaltered from the original.

However, Abingo.score_conversion! substituted spaces with underscores
before checking if the test existed in participating tests. Hence,
conversions would never be scored if the test name contained spaces.

I considered the possibility that test names were not intended to have
spaces, but couldn't find any documentation for this, and the
substitution code suggests otherwise.

This fixes the problem by only substituting the spaces in the test name
when generating the cache key in the score_conversion! function. A unit
test was modified to test the test with spaces scenario.